requirements.txt
playwright==1.48.0
pytest==8.3.3
pytest-playwright==0.5.2
pytest-html==4.1.1
pytest-xdist==3.6.1
allure-pytest==2.13.5
python-dotenv==1.0.1



